# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v2/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Events based on User Roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get All Events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.List"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetAllEventsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/events/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Event and Instances by Event Code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event by Event Code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "object that needs to be added",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ListWithDetail"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetInstancesByEventCodeResponse"
                                            }
                                        },
                                        "details": {
                                            "$ref": "#/definitions/models.GetEventByCodeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v2/internal/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create event with the instances/sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create Event",
                "parameters": [
                    {
                        "description": "User object that needs to be added",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.CreateEventResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "instances": {
                                            "$ref": "#/definitions/models.CreateInstanceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateEventRequest": {
            "type": "object",
            "required": [
                "allowedCampuses",
                "allowedFor",
                "allowedRoles",
                "allowedUsers",
                "instances",
                "locationName",
                "locationType",
                "name"
            ],
            "properties": {
                "allowedCampuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedFor": {
                    "type": "string",
                    "enum": [
                        "public",
                        "private"
                    ]
                },
                "allowedRoles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedUsers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "eventEndAt": {
                    "type": "string"
                },
                "eventStartAt": {
                    "type": "string"
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateInstanceRequest"
                    }
                },
                "isRecurring": {
                    "type": "boolean"
                },
                "locationName": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string",
                    "enum": [
                        "online",
                        "onsite",
                        "hybrid"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "recurrence": {
                    "type": "string"
                },
                "registerEndAt": {
                    "type": "string"
                },
                "registerStartAt": {
                    "type": "string"
                },
                "termsAndConditions": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.CreateEventResponse": {
            "type": "object",
            "required": [
                "instances"
            ],
            "properties": {
                "allowedCampuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedFor": {
                    "type": "string",
                    "example": "public"
                },
                "allowedRoles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedUsers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "code": {
                    "type": "string",
                    "example": "bhfe382"
                },
                "description": {
                    "type": "string",
                    "example": "This event blabla"
                },
                "eventEndAt": {
                    "type": "string",
                    "example": ""
                },
                "eventStartAt": {
                    "type": "string",
                    "example": ""
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateInstanceResponse"
                    }
                },
                "isRecurring": {
                    "type": "boolean",
                    "example": true
                },
                "locationName": {
                    "type": "string",
                    "example": "PIOT 6 Lt. 6"
                },
                "locationType": {
                    "type": "string",
                    "example": "offline"
                },
                "recurrence": {
                    "type": "string",
                    "example": "monthly"
                },
                "registerEndAt": {
                    "type": "string",
                    "example": ""
                },
                "registerStartAt": {
                    "type": "string",
                    "example": ""
                },
                "status": {
                    "type": "string",
                    "example": "available"
                },
                "termsAndConditions": {
                    "type": "string",
                    "example": "This event blabla"
                },
                "title": {
                    "type": "string",
                    "example": "Homebase"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "event"
                }
            }
        },
...
        "models.CreateInstanceRequest": {
            "type": "object",
            "required": [
                "allowVerifyAt",
                "disallowVerifyAt",
                "instanceEndAt",
                "instanceStartAt",
                "locationName",
                "locationType",
                "registerEndAt",
                "registerStartAt",
                "title"
            ],
            "properties": {
                "allowVerifyAt": {
                    "type": "string"
                },
                "checkType": {
                    "type": "string",
                    "enum": [
                        "check-in",
                        "check-out",
                        "both",
                        "none"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "disallowVerifyAt": {
                    "type": "string"
                },
                "instanceEndAt": {
                    "type": "string"
                },
                "instanceStartAt": {
                    "type": "string"
                },
                "isOnePerAccount": {
                    "type": "boolean"
                },
                "isOnePerTicket": {
                    "type": "boolean"
                },
                "locationName": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string",
                    "enum": [
                        "online",
                        "onsite",
                        "hybrid"
                    ]
                },
                "maxPerTransaction": {
                    "type": "integer"
                },
                "registerEndAt": {
                    "type": "string"
                },
                "registerFlow": {
                    "type": "string",
                    "enum": [
                        "personal-qr",
                        "event-qr",
                        "both-qr",
                        "none"
                    ]
                },
                "registerStartAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalSeats": {
                    "type": "integer"
                }
            }
        },
        "models.CreateInstanceResponse": {
            "type": "object",
            "properties": {
                "allowVerifyAt": {
                    "type": "string"
                },
                "checkType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "disallowVerifyAt": {
                    "type": "string"
                },
                "eventCode": {
                    "type": "string"
                },
                "instanceCode": {
                    "type": "string"
                },
                "instanceEndAt": {
                    "type": "string"
                },
                "instanceStartAt": {
                    "type": "string"
                },
                "isOnePerAccount": {
                    "type": "boolean"
                },
                "isOnePerTicket": {
                    "type": "boolean"
                },
                "locationName": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string"
                },
                "maxPerTransaction": {
                    "type": "integer"
                },
                "registerEndAt": {
                    "type": "string"
                },
                "registerFlow": {
                    "type": "string"
                },
                "registerStartAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "title": {
                    "type": "string"
                },
                "totalSeats": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ErrorValidateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "accountNumber_required"
                },
                "field": {
                    "type": "string",
                    "example": "MISSING_FIELD"
                },
                "message": {
                    "type": "string",
                    "example": "field is missing"
                }
            }
        },
        "models.GetAllEventsResponse": {
            "type": "object",
            "properties": {
                "allowedCampuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedFor": {
                    "type": "string",
                    "example": "public"
                },
                "allowedRoles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedUsers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "availabilityStatus": {
                    "type": "string",
                    "example": "available"
                },
                "code": {
                    "type": "string",
                    "example": "2024-HOMEBASE"
                },
                "eventEndAt": {
                    "type": "string",
                    "example": ""
                },
                "eventStartAt": {
                    "type": "string",
                    "example": ""
                },
                "imagesLinks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isRecurring": {
                    "type": "boolean",
                    "example": true
                },
                "locationType": {
                    "type": "string",
                    "example": "offline"
                },
                "recurrence": {
                    "type": "string",
                    "example": "monthly"
                },
                "registerEndAt": {
                    "type": "string",
                    "example": ""
                },
                "registerStartAt": {
                    "type": "string",
                    "example": ""
                },
                "title": {
                    "type": "string",
                    "example": "Homebase"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "totalRemainingSeats": {
                    "type": "integer",
                    "example": 2
                },
                "type": {
                    "type": "string",
                    "example": "Event"
                }
            }
        },
...
        "models.GetEventByCodeResponse": {
            "type": "object",
            "properties": {
                "allowedCampuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedFor": {
                    "type": "string",
                    "example": "public"
                },
                "allowedRoles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedUsers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "availabilityStatus": {
                    "type": "string",
                    "example": "available"
                },
                "code": {
                    "type": "string",
                    "example": "bhfe382"
                },
                "description": {
                    "type": "string",
                    "example": "This event blabla"
                },
                "eventEndAt": {
                    "type": "string",
                    "example": ""
                },
                "eventStartAt": {
                    "type": "string",
                    "example": ""
                },
                "imageLinks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetInstancesByEventCodeResponse"
                    }
                },
                "isRecurring": {
                    "type": "boolean",
                    "example": true
                },
                "locationName": {
                    "type": "string",
                    "example": "PIOT 6 Lt. 6"
                },
                "locationType": {
                    "type": "string",
                    "example": "offline"
                },
                "recurrence": {
                    "type": "string",
                    "example": "monthly"
                },
                "registerEndAt": {
                    "type": "string",
                    "example": ""
                },
                "registerStartAt": {
                    "type": "string",
                    "example": ""
                },
                "termsAndConditions": {
                    "type": "string",
                    "example": "This event blabla"
                },
                "title": {
                    "type": "string",
                    "example": "Homebase"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "event"
                }
            }
        },
...
        "models.GetInstancesByEventCodeResponse": {
            "type": "object",
            "properties": {
                "LocationName": {
                    "type": "string",
                    "example": "PIOT 6 Lt. 6"
                },
                "allowVerifyAt": {
                    "type": "string",
                    "example": ""
                },
                "availabilityStatus": {
                    "type": "string",
                    "example": "available"
                },
                "bookedSeats": {
                    "type": "integer",
                    "example": 0
                },
                "checkType": {
                    "type": "string"
                },
                "code": {
                    "type": "string",
                    "example": "2024-HOMEBASE"
                },
                "description": {
                    "type": "string",
                    "example": "Homebase"
                },
                "disallowVerifyAt": {
                    "type": "string",
                    "example": ""
                },
                "instanceEndAt": {
                    "type": "string",
                    "example": ""
                },
                "instanceStartAt": {
                    "type": "string",
                    "example": ""
                },
                "isOnePerAccount": {
                    "type": "boolean"
                },
                "isOnePerTicket": {
                    "type": "boolean"
                },
                "locationType": {
                    "type": "string",
                    "example": "offline"
                },
                "maxPerTransaction": {
                    "type": "integer"
                },
                "registerEndAt": {
                    "type": "string",
                    "example": ""
                },
                "registerFlow": {
                    "type": "string"
                },
                "registerStartAt": {
                    "type": "string",
                    "example": ""
                },
                "title": {
                    "type": "string",
                    "example": "Homebase"
                },
                "totalRemainingSeats": {
                    "type": "integer",
                    "example": 0
                },
                "totalSeats": {
                    "type": "integer",
                    "example": 0
                },
                "type": {
                    "type": "string",
                    "example": "eventInstance"
                }
            }
        },
...
        "models.List": {
            "type": "object",
            "properties": {
                "data": {},
                "totalRows": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "example": "collection"
                }
            }
        },
        "models.ListWithDetail": {
            "type": "object",
            "properties": {
                "data": {},
                "details": {},
                "totalRows": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "example": "collection"
                }
            }
        },
...
        "models.Metadata": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}